// src/api/planets/getPage/types.ts
type PlanetsRaw = {
  name: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
};

export type Response = {
  count: number;
  next: string | null;
  previous: string | null;
  results: PlanetsRaw[];
};

export type Result = {
  count: number;
  next: string | null;
  previous: string | null;
  results: Planet[];
};

export type Planet = {
  name: string;
  rotationPeriod: string;
  orbitalPeriod: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surfaceWater: string;
  population: string;
  residents: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
};